Improve valuesToMatrix, fix supposed bug
There is no need to fill the matrix in a loop. A call
to matrix() already does the trick. It is important
to set byrow = TRUE so that the values are arranged
row by row instead of column by column. The loop
filled the matrix from the bottom row to the top row.
This is achieved now by indexing the matrix rows
with a reversed vector of 1:nrow.
This version removes the (assumed) error that the
last row of the matrix was not filled (because
ny -j never gets ny) and that the last values
(appearing in the first row) were lost (because
ny - j is 0 for j = ny and unfortunately assiging
values to matrix row 0 does not result in an error).